feat: POC of Clean Architecture
