+++ apps/dokploy/pages/dashboard/projects.tsx
@@ -4,8 +4,8 @@ import type { GetServerSidePropsContext } from "next";
 import dynamic from "next/dynamic";
 import type { ReactElement } from "react";
 import superjson from "superjson";
-import { ShowProjects } from "@/components/dashboard/projects/show";
 import { DashboardLayout } from "@/components/layouts/dashboard-layout";
+import { ShowProjects } from "@/features/projects/ui/containers/show-projects";
 import { appRouter } from "@/server/api/root";
 import { api } from "@/utils/api";
